pyproject.toml
[project]
name = "redeye"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "karl3408", email = "karl34408@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-multipart",
    "requests",
    "python-dotenv",
    "datasets",
    "transformers",
    "scikit-learn",
    "evaluate",
    "accelerate",
    "pandas",
    "torch>=2.5.1",
    "torchvision",
    "torchaudio",
    "tiktoken",
    "motor",
    "pymongo",
    "python-owasp-zap-v2-4",
    # Stable LangChain 0.2.x Stack
    "langchain==0.2.16",
    "langchain-community==0.2.16",
    "langchain-core==0.2.38",
    "langchain-openai==0.1.23",
    "langchain-mongodb==0.1.6",
    "langchain-text-splitters==0.2.4",
]

[project.scripts]
redeye = "redeye:main"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch" }]
torchvision = [{ index = "pytorch" }]
torchaudio = [{ index = "pytorch" }]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
